Add WithTrailingApySchema to separate file
